training/ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
torch==2.5.1
torchaudio==2.5.1
transformers==4.47.1
datasets==3.2.0
evaluate==0.4.3
jiwer==3.1.0
librosa==0.10.2
soundfile==0.13.1
numpy==1.26.4
scipy==1.14.1
pydub==0.25.1
onnx==1.17.0
onnxruntime==1.20.1
optimum==1.24.0
accelerate==1.2.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
redis==5.2.1
celery==5.4.0
